Restructure services pages and components.
